make cryptoRepositoryFactory a simple function
